+++ mfit/documentation/release_notes.md
@@ -2,8 +2,9 @@
 
 ## v1.1.0 (Pending)
 ##### Features
-- Added enumerations for EventTopics and EventStates (`ME-617`).
-- Added interface definitions for Events and LogConstructors (`ME-617`).
+- Added Event model (`ME-617`).
+- Added EventTopic and EventState enumerations (`ME-617`).
+- Added Event and LogConstructor interface definitions (`ME-617`).
 
 ## v1.0.1 (2017-03-31)
 ##### Features